move dependencies to gemspec
Note: I did remove the dependency on `rails >= 4` in favor of a weaker (>= 3) dependency on `activesupport`. it seems like `rails` isn't really required for this library, and I think the only thing from `activesupport` is being used for is some `deep_merge` and `try` statements, so I'm only requiring its core extensions.